refactor!: introduce core array operation traits
